+++ queue.h
@@ -6,7 +6,7 @@
 #include <stdint.h>
 #include  <stdlib.h> // malloc
 // #define DEBUG
-typedef struct {int16_t * buffer;
+typedef struct {uint16_t * buffer;
                 uint16_t head;
                 uint16_t tail;
                 uint16_t max_index;
@@ -14,7 +14,7 @@ typedef struct {int16_t * buffer;
 // ACQUIRE BUFFER USING MALLOC AT INITIALIZATION.
 
 //API
-void init_queue(queue_t * ptr_to_existing_q, int max_items);
+void init_queue(queue_t * ptr_to_existing_q, uint16_t max_items);
 bool read_q(queue_t *q0, int16_t *datum);
 bool write_q(queue_t *q0, int16_t datum);
 //queue_t x;
